move view logic into a separate component
